Separate out clearing of output and creation of output directory
This was causing problems on some platforms, as sometimes the build path
is a file, and sometimes it's a directory. The old code always created a
directory, which Unity wouldn't like when it came to build in that
location. Instead, the parent directory is created so Unity has somewhere
to build into, and this is always done rather than only when the user has
requested a clean build.